added intro and scholar branches
