+++ 2023 Draft/final/master_copy.py
@@ -5,116 +5,6 @@
 # Talos Principle, all rights are reserved for the owner.
 
 #######################################################################################################
-	
-
-########################################################################################################
-
-# This Func might be able to reduced by using inventory.index(item).
-# However, further research shows that .index() assumes the item is there,
-# so if the item is not found a ValueError will occur. So the current
-# code is sufficent. 
-
-# Inventory check to show hidden blacksmith option
-def item_check(item):
-	check = 0
-	for line in inventory:
-		if line == str(item):
-			check += 1
-	return check
-
-################################################################################################
-
-
-
-#################################################################################################
-# This section is for scholar storyline 
-
-# Helper function to pause text as it is printed
-def print_with_pause(text_list):
-    for text, pause in text_list:
-        print(text)
-        time.sleep(pause)
-
-# This is the intro Scholar text for the scholar path. 
-# Potentially putting this in a sepereate text file that may be a better method of printing with pause
-def scholar_text():
-    scholar_intro_text = [
-        ('XXXXXXXXXXXXXXXXXXXXXXXX', .1),
-        ('XXTHE BLOOD OF SCHOLARSX', .1),
-        ('XXXXXXXXXXXXXXXXXXXXXXXX', .1),
-        ('XSHALL OUTWEIGTHXXXXXXXX', 2),
-        ('XXXXXXXXXXXXXXXXXXXXXXXX', .1),
-        ('XXTHE BLOOD OF MARTYRSXX', .1),
-        ('XXXXXXXXXXXXXXXXXXXXXXXX', .1),
-        ('', 0),
-        ('You are Arkady', 0.5),
-        ('        the Scholar', 0.5),
-        ('        the Chronicler', 0.5),
-        ('        the Sage', 0.5),
-        ('', 0),
-        ('XX The world has been consumed by ice. XXXX', 1),
-        ('XX In the Last Monastery XXXXXXXXXXXXXXXXXX', 1),
-        ('XX of the land of Faith, XXXXXXXXXXXXXXXXXX', 1),
-        ('XX you complete the thoughts of the dead XX', 1),
-        ('XXXXXXX that God might know who they were X,', 1),
-        ('', 0),
-        ('The hazy ghost of Weariness is always upon you, breaking your will. ', .1),
-        ('You fear that you may fail before the Great Work is finished.', .1),
-        ('', 0)
-    ]
-    print_with_pause(scholar_intro_text)
-
-'''
-Use case:
-scholar_text()
-
-Result:
-Prints out the text with pauses at specific places.
-'''
-
-
-# Below is the obsolete method of printing the scholar intro text. 
-'''
-# This is the intro Scholar text for the scholar path. 
-# Potentially putting this in a sepereate text file that may be a better method of printing with pause
-def scholar_text():
-        print('XXXXXXXXXXXXXXXXXXXXXXXX')
-        print('XXTHE BLOOD OF SCHOLARSX')
-        print('XXXXXXXXXXXXXXXXXXXXXXXX')
-        time.sleep(2)
-        print('XSHALL OUTWEIGTHXXXXXXXX')
-        print('XXXXXXXXXXXXXXXXXXXXXXXX')
-        print('XXTHE BLOOD OF MARTYRSXX')
-        print('XXXXXXXXXXXXXXXXXXXXXXXX')
-        print()
-        time.sleep(2)
-        print('You are Arkady')
-        time.sleep(.5)
-        print('        the Scholar')
-        time.sleep(.5)
-        print('        the Chronicler')
-        time.sleep(.5)
-        print('        the Sage')
-        print('')
-        time.sleep(1)
-        print('XX The world has been consumed by ice. XXXX')
-        time.sleep(1)
-        print('XX In the Last Monastery XXXXXXXXXXXXXXXXXX')
-        time.sleep(1)
-        print('XX of the land of Faith, XXXXXXXXXXXXXXXXXX')
-        time.sleep(1)
-        print('XX you complete the thoughts of the dead XX')
-        time.sleep(1)
-        print('XXXXXXX that God might know who they were X,')
-        time.sleep(1)
-        print()
-        print('The hazy ghost of Weariness is always upon you, breaking your will. ')
-        print('You fear that you may fail before the Great Work is finished.')
-        print()
-'''
-########################################################################################################
-
-
 
 # Func goes through y/n work loop
 def work():
