+++ old/README.md
@@ -1,0 +1,5 @@
+This is a recreation of a mini-game found within the video game "The Talos Principle" to learn Python and common coding techniques. The purpose of this project is to demostrate the an evolution of my coding skills. 
+I started this project when I was early in my programing career and currently this code is full of inefficiencies and poor coding practices. By the end of this project I will demostrate how effecient and streamline
+python coding practices. 
+
+This is intended for recreational and learning purposes. I do not have any rights to this game.
