+++ old/Scholar_Dialog.py
