+++ old/Scholar_Draft.py
