+++ old/intro.txt
@@ -7,3 +7,4 @@ XX
    Your name in that time was ANTHROPOS.
 XX
 XX Cast your mind back to that time now. XX
+
