+++ old/intro2.txt
@@ -12,3 +12,4 @@ But strife came to the city, and with strife came sickness.
 Thus you, the giant ANTHROPOS, fell into a feverish sleep
    and the city was torn apart
    XX for the city is the PEOPLE, and people are the CITY. XX
+
