+++ old/master_copy.py
